+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.23.3] 2026/02/03
+
+### Changed
+- Extracted reward level redemption logic into `Rewards::RedeemLevel`.
+- Extracted `redeem_any` logic into `Rewards::RedeemAny`.
+- Extracted `redeem_task` logic into `Rewards::RedeemTask`.
+- Extracted rewards index aggregation into `Rewards::IndexData`.
+
 ## [1.23.2] 2026/02/03
 
 ### Changed